Adding methods for Penelope functionality
